localize integration mutations
