+++ src/pages/IntegrationCallbackPage/IntegrationCallbackPage.tsx
@@ -1,11 +1,13 @@
+// TODO move this to pages/CircleAdminPage or features/integrations
+
 import { FC, useCallback, useEffect, useMemo, useState } from 'react';
 
-import * as mutations from 'lib/gql/mutations';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import { Typography } from '@material-ui/core';
 
 import { LoadingModal } from 'components';
+import { createCircleIntegration } from 'pages/CircleAdminPage/mutations';
 import { useSelectedCircle } from 'recoilState';
 import { paths } from 'routes/paths';
 import { Button } from 'ui';
@@ -78,7 +80,7 @@ export const IntegrationCallbackPage: FC = () => {
     if (integration) {
       const data = integration.create(params);
       try {
-        await mutations.createCircleIntegration(
+        await createCircleIntegration(
           circleId,
           integration.name,
           data.integrationName,
